requirements.txt
streamlit
langchain
langchain-community
langchain-groq
cohere
faiss-cpu
python-dotenv
beautifulsoup4